vcl: a way to create data for BinaryDataContainer without copying
Construstor for BinaryDataContainer taking unique_ptr<sal_uInt8>
represents a way how to create and fill the data for the container
and make sure that the container is the only one that is managing
the data after construction.
Generally we don't want to give the access to the internal
shared_ptr to the outside, so with the unique_ptr we make sure
that there is no outside references present (which would be the
case if the we would take shared_ptr as the constructor) and the
unique_ptr is cleared (moved) after construction.

Change-Id: I123114c8bee92e342740d94a784b2c16e2564528